Neonatal Documentation Update
I'm updating the core neonatal documentation with a few tweaks to reflect the addition of a new year of data. Most of it is minor changes. 

I've also added a "checklist" at the bottom of the Downloading Data Process section, as that was helpful for me as I went through the process myself. It takes the narrative style of the original documentation and puts it in list form. Feel free to include this in the document or not - it's a repetition of information that's already there, just organized differently.