Calculate data quality flag using total population
Use total population in data quality flag instead of homeowners

This is because we expect that we will have people of color shut out of homeownership opportunities, and if they make up a large part of a community, but have a small share of the collective home value, that will be a strong quality finding.

Conversely, data quality is lower if the group has a small share of the collective home values but they already make up such a small share of the community's population.